refactor: page.tsxを薄いラッパーに変更しロジックをfeature層に移動
CodeRabbitのレビュー指摘に対応。
- loadDisclosureDataローダーをfeature層に新設
- page.tsxはパラメータ取得とコンポーネント呼び出しのみに
