+++ web/src/app/bills/[id]/interview/disclosure/page.tsx
@@ -2,11 +2,8 @@ import type { Metadata } from "next";
 import { notFound } from "next/navigation";
 import { getBillById } from "@/features/bills/server/loaders/get-bill-by-id";
 import { getInterviewConfig } from "@/features/interview-config/server/loaders/get-interview-config";
-import { getInterviewQuestions } from "@/features/interview-config/server/loaders/get-interview-questions";
+import { loadDisclosureData } from "@/features/interview-config/server/loaders/load-disclosure-data";
 import { InterviewDisclosurePage } from "@/features/interview-config/server/components/interview-disclosure-page";
-import { buildBulkModeSystemPrompt } from "@/features/interview-session/shared/utils/interview-logic/bulk-mode";
-import { buildLoopModeSystemPrompt } from "@/features/interview-session/shared/utils/interview-logic/loop-mode";
-import { buildSummarySystemPrompt } from "@/features/interview-session/shared/utils/build-summary-system-prompt";
 
 interface DisclosurePageProps {
   params: Promise<{
@@ -41,43 +38,11 @@ export default async function DisclosurePage({ params }: DisclosurePageProps) {
     getInterviewConfig(id),
   ]);
 
-  if (!bill) {
+  if (!bill || !interviewConfig) {
     notFound();
   }
 
-  if (!interviewConfig) {
-    notFound();
-  }
+  const disclosureData = await loadDisclosureData(bill, interviewConfig);
 
-  const questions = await getInterviewQuestions(interviewConfig.id);
-
-  const mode = interviewConfig.mode ?? "loop";
-  const buildSystemPrompt =
-    mode === "bulk" ? buildBulkModeSystemPrompt : buildLoopModeSystemPrompt;
-
-  const systemPrompt = buildSystemPrompt({
-    bill,
-    interviewConfig,
-    questions,
-    currentStage: "chat",
-    askedQuestionIds: new Set(),
-    remainingMinutes: null,
-  });
-
-  const summaryPrompt = buildSummarySystemPrompt({
-    bill,
-    interviewConfig,
-    messages: [],
-  });
-
-  return (
-    <InterviewDisclosurePage
-      billName={bill.name}
-      billTitle={bill.bill_content?.title ?? bill.name}
-      interviewConfig={interviewConfig}
-      questions={questions}
-      systemPrompt={systemPrompt}
-      summaryPrompt={summaryPrompt}
-    />
-  );
+  return <InterviewDisclosurePage {...disclosureData} />;
 }
